+++ tests/specs/class/memberReturnThis.d.ts
@@ -1,0 +1,7 @@
+export interface MyObject {
+    instance: () => this;
+
+    instance1: (a : Boolean) => this;
+
+    instance2: (a : Boolean, b : number) => this;
+}